+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/implementation/ProtectionContainersClientImpl.java
@@ -277,14 +277,7 @@ public final class ProtectionContainersClientImpl implements ProtectionContainer
     private Mono<ProtectionContainerResourceInner> getAsync(
         String vaultName, String resourceGroupName, String fabricName, String containerName) {
         return getWithResponseAsync(vaultName, resourceGroupName, fabricName, containerName)
-            .flatMap(
-                (Response<ProtectionContainerResourceInner> res) -> {
-                    if (res.getValue() != null) {
-                        return Mono.just(res.getValue());
-                    } else {
-                        return Mono.empty();
-                    }
-                });
+            .flatMap(res -> Mono.justOrEmpty(res.getValue()));
     }
 
     /**
@@ -487,14 +480,7 @@ public final class ProtectionContainersClientImpl implements ProtectionContainer
         String containerName,
         ProtectionContainerResourceInner parameters) {
         return registerWithResponseAsync(vaultName, resourceGroupName, fabricName, containerName, parameters)
-            .flatMap(
-                (Response<ProtectionContainerResourceInner> res) -> {
-                    if (res.getValue() != null) {
-                        return Mono.just(res.getValue());
-                    } else {
-                        return Mono.empty();
-                    }
-                });
+            .flatMap(res -> Mono.justOrEmpty(res.getValue()));
     }
 
     /**
@@ -684,7 +670,7 @@ public final class ProtectionContainersClientImpl implements ProtectionContainer
     private Mono<Void> unregisterAsync(
         String vaultName, String resourceGroupName, String fabricName, String containerName) {
         return unregisterWithResponseAsync(vaultName, resourceGroupName, fabricName, containerName)
-            .flatMap((Response<Void> res) -> Mono.empty());
+            .flatMap(ignored -> Mono.empty());
     }
 
     /**
@@ -866,7 +852,7 @@ public final class ProtectionContainersClientImpl implements ProtectionContainer
     private Mono<Void> inquireAsync(
         String vaultName, String resourceGroupName, String fabricName, String containerName, String filter) {
         return inquireWithResponseAsync(vaultName, resourceGroupName, fabricName, containerName, filter)
-            .flatMap((Response<Void> res) -> Mono.empty());
+            .flatMap(ignored -> Mono.empty());
     }
 
     /**
@@ -886,7 +872,7 @@ public final class ProtectionContainersClientImpl implements ProtectionContainer
         String vaultName, String resourceGroupName, String fabricName, String containerName) {
         final String filter = null;
         return inquireWithResponseAsync(vaultName, resourceGroupName, fabricName, containerName, filter)
-            .flatMap((Response<Void> res) -> Mono.empty());
+            .flatMap(ignored -> Mono.empty());
     }
 
     /**
@@ -1058,7 +1044,7 @@ public final class ProtectionContainersClientImpl implements ProtectionContainer
     @ServiceMethod(returns = ReturnType.SINGLE)
     private Mono<Void> refreshAsync(String vaultName, String resourceGroupName, String fabricName, String filter) {
         return refreshWithResponseAsync(vaultName, resourceGroupName, fabricName, filter)
-            .flatMap((Response<Void> res) -> Mono.empty());
+            .flatMap(ignored -> Mono.empty());
     }
 
     /**
@@ -1077,7 +1063,7 @@ public final class ProtectionContainersClientImpl implements ProtectionContainer
     private Mono<Void> refreshAsync(String vaultName, String resourceGroupName, String fabricName) {
         final String filter = null;
         return refreshWithResponseAsync(vaultName, resourceGroupName, fabricName, filter)
-            .flatMap((Response<Void> res) -> Mono.empty());
+            .flatMap(ignored -> Mono.empty());
     }
 
     /**