+++ sdk/resourcemanager/azure-resourcemanager-compute/src/main/java/com/azure/resourcemanager/compute/implementation/SharedGalleryImagesClientImpl.java
@@ -438,14 +438,7 @@ public final class SharedGalleryImagesClientImpl implements SharedGalleryImagesC
     @ServiceMethod(returns = ReturnType.SINGLE)
     public Mono<SharedGalleryImageInner> getAsync(String location, String galleryUniqueName, String galleryImageName) {
         return getWithResponseAsync(location, galleryUniqueName, galleryImageName)
-            .flatMap(
-                (Response<SharedGalleryImageInner> res) -> {
-                    if (res.getValue() != null) {
-                        return Mono.just(res.getValue());
-                    } else {
-                        return Mono.empty();
-                    }
-                });
+            .flatMap(res -> Mono.justOrEmpty(res.getValue()));
     }
 
     /**