+++ sdk/resourcemanager/azure-resourcemanager-compute/src/main/java/com/azure/resourcemanager/compute/implementation/SshPublicKeysClientImpl.java
@@ -637,14 +637,7 @@ public final class SshPublicKeysClientImpl
     public Mono<SshPublicKeyResourceInner> createAsync(
         String resourceGroupName, String sshPublicKeyName, SshPublicKeyResourceInner parameters) {
         return createWithResponseAsync(resourceGroupName, sshPublicKeyName, parameters)
-            .flatMap(
-                (Response<SshPublicKeyResourceInner> res) -> {
-                    if (res.getValue() != null) {
-                        return Mono.just(res.getValue());
-                    } else {
-                        return Mono.empty();
-                    }
-                });
+            .flatMap(res -> Mono.justOrEmpty(res.getValue()));
     }
 
     /**
@@ -811,14 +804,7 @@ public final class SshPublicKeysClientImpl
     public Mono<SshPublicKeyResourceInner> updateAsync(
         String resourceGroupName, String sshPublicKeyName, SshPublicKeyUpdateResource parameters) {
         return updateWithResponseAsync(resourceGroupName, sshPublicKeyName, parameters)
-            .flatMap(
-                (Response<SshPublicKeyResourceInner> res) -> {
-                    if (res.getValue() != null) {
-                        return Mono.just(res.getValue());
-                    } else {
-                        return Mono.empty();
-                    }
-                });
+            .flatMap(res -> Mono.justOrEmpty(res.getValue()));
     }
 
     /**
@@ -965,8 +951,7 @@ public final class SshPublicKeysClientImpl
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     public Mono<Void> deleteAsync(String resourceGroupName, String sshPublicKeyName) {
-        return deleteWithResponseAsync(resourceGroupName, sshPublicKeyName)
-            .flatMap((Response<Void> res) -> Mono.empty());
+        return deleteWithResponseAsync(resourceGroupName, sshPublicKeyName).flatMap(ignored -> Mono.empty());
     }
 
     /**
@@ -1112,14 +1097,7 @@ public final class SshPublicKeysClientImpl
     @ServiceMethod(returns = ReturnType.SINGLE)
     public Mono<SshPublicKeyResourceInner> getByResourceGroupAsync(String resourceGroupName, String sshPublicKeyName) {
         return getByResourceGroupWithResponseAsync(resourceGroupName, sshPublicKeyName)
-            .flatMap(
-                (Response<SshPublicKeyResourceInner> res) -> {
-                    if (res.getValue() != null) {
-                        return Mono.just(res.getValue());
-                    } else {
-                        return Mono.empty();
-                    }
-                });
+            .flatMap(res -> Mono.justOrEmpty(res.getValue()));
     }
 
     /**
@@ -1271,14 +1249,7 @@ public final class SshPublicKeysClientImpl
     public Mono<SshPublicKeyGenerateKeyPairResultInner> generateKeyPairAsync(
         String resourceGroupName, String sshPublicKeyName) {
         return generateKeyPairWithResponseAsync(resourceGroupName, sshPublicKeyName)
-            .flatMap(
-                (Response<SshPublicKeyGenerateKeyPairResultInner> res) -> {
-                    if (res.getValue() != null) {
-                        return Mono.just(res.getValue());
-                    } else {
-                        return Mono.empty();
-                    }
-                });
+            .flatMap(res -> Mono.justOrEmpty(res.getValue()));
     }
 
     /**