+++ sdk/core/azure-core-http-netty/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.13.0-beta.1 (Unreleased)
 
 ### Features Added
+- Added specialized consumption for `HttpRequest.getBodyAsBinaryData()`.
 
 ### Breaking Changes