+++ sdk/core/azure-core-http-netty/CHANGELOG.md
@@ -3,13 +3,14 @@
 ## 1.13.0-beta.1 (Unreleased)
 
 ### Features Added
-- Added specialized consumption for `HttpRequest.getBodyAsBinaryData()`.
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+- Added specialized consumption for `HttpRequest.getBodyAsBinaryData()`.
 
 ## 1.12.0 (2022-05-06)